merkle: collapse `common` module into `mod.rs`
