+++ src/merkle/imr.rs
@@ -17,8 +17,8 @@ use crate::core::Value;
 use crate::impl_midstate_wrapper;
 use crate::jet::Jet;
 use crate::merkle::cmr::Cmr;
-use crate::merkle::common::{CommitMerkleRoot, MerkleRoot};
 use crate::merkle::tmr::Tmr;
+use crate::merkle::{CommitMerkleRoot, MerkleRoot};
 use crate::types::arrow::FinalArrow;
 use bitcoin_hashes::sha256::Midstate;
